redis序列化变更为Jackson，更稳定与安全

+++ lucky-common/lucky-common-redis/src/main/java/com/lucky/common/redis/utils/RedisCache.java
@@ -43,7 +43,7 @@ public class RedisCache {
      * @param timeout  时间
      * @param timeUnit 时间颗粒度
      */
-    public <T> void setCacheObject(final String key, final T value, final Integer timeout, final TimeUnit timeUnit) {
+    public <T> void setCacheObject(final String key, final T value, final long timeout, final TimeUnit timeUnit) {
         redisTemplate.opsForValue().set(key, value, timeout, timeUnit);
     }
